(Collatinus) Fixed an issue with collatinus format data
Sometime, not so rarely, the models are defined with desinences such as :

`des:1-2:a`

Where 1-2 means desinence 1 and 2. This often happens for ablative (only once `ibus` in the line). What this is meant to do is that, if you have finished dispatching
all desinences, you are repeating the last one you encountered.