Update collatinus decliner (#8)
* (Collatinus) Updated decliner data, still needs some works (check ToDo)

* (Collatinus) Updated to deal with variations on nominative

* (Collatinus) Went another way for dealing with multiple radicals and forms

* (Collatinus) Added map for helping user find their right lemma

* (Collatinus) Fixed an issue with collatinus format data

Sometime, not so rarely, the models are defined with desinences such as :

`des:1-2:a`

Where 1-2 means desinence 1 and 2. This often happens for ablative (only once `ibus` in the line). What this is meant to do is that, if you have finished dispatching
all desinences, you are repeating the last one you encountered.